Bump `Pillow`
This saves us a headache trying to build a version of Pillow older than
that in current Nixpkgs.

The Nixpkgs derivation for the newer Pillow rewrites setup.py with
libavif paths, but the AVIF entry is not there in the older version, and
so the substitution fails, breaking the build. We can either bump Pillow
in Bookwyrm, or override the whole `preConfigure` here, and bumping
Pillow is the more expedient option.